working with logic implementation

+++ symtest.Common/Models/HttpRequestTemplate.cs
@@ -9,8 +9,8 @@ namespace symtest.Common.Models
         public HttpMethod Method { get; set; }
         
         public int Density { get; set; }
-        public float Duration { get; set; }
-        public float Distribution { get; set; }
+        public int Duration { get; set; }
+        public double Distribution { get; set; }
         
         public Dictionary<string, string> Headers { get; set; }
     }
